firstproject_root/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.8
sqlparse==0.4.4
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==9.0.1


